websockets-video-controls-app/requirements.txt
microdot
RPi.GPIO
websockets
asyncio
opencv-python
aiohttp
depthai
pyserial>=3.5,<4
